src/end_device/visual-predictor/requirements.txt
setuptools
Flask
mysql-connector-python
pyyaml
statsmodels
sqlalchemy